a last test on conditionnal plumed loading
